Refactored Sanity files
