+++ app/src/lib/sanity/queries/event/getEvents.ts
@@ -1,0 +1,13 @@
+const getEvents = `
+*[_type == "event"][] | order(date desc){
+  title,
+  coverImage{asset->{url, metadata{lqip, dimensions}}},
+  tag[]->{title},
+  date,
+  content[]{...,
+    _type == 'image' => @{asset->{url, metadata{lqip, dimensions}}},
+  }
+}
+`;
+
+export default getEvents;
