+++ app/src/lib/sanity/queries/index.ts
@@ -1,0 +1,10 @@
+import getMetaData from './metaData/getMetaData';
+import type MetaData from './metaData/metaData';
+import getPosts from './post/getPosts';
+import getPost from './post/getPost';
+import type Post from './post/post';
+import getPage from './page/getPage';
+import type SanityPage from './page/page';
+
+export { getMetaData, getPost, getPosts, getPage };
+export { MetaData, Post, SanityPage };
