+++ app/src/lib/sanity/queries/page/objects/cta.d.ts
