+++ app/src/lib/sanity/queries/page/objects/header.d.ts
