+++ app/src/lib/sanity/queries/page/objects/hero.d.ts
