+++ app/src/lib/sanity/queries/page/objects/mediaGallery.d.ts
