+++ app/src/lib/sanity/queries/page/objects/team.d.ts
