+++ app/src/lib/sanity/queries/post/post.d.ts
@@ -50,3 +50,5 @@ export type Child = {
 	_type: string;
 	marks: any[];
 };
+
+export default Post;
