--- app/src/lib/sanity/query.d.ts
@@ -1,6 +1,0 @@
-import type { Post } from './types/post'
-
-export type ReqGetPosts = {
-    posts: Post[],
-    total: number
-} | undefined
