--- app/src/lib/sanity/types/image.d.ts
@@ -1,12 +1,0 @@
-// Generated by https://quicktype.io
-
-export interface Image {
-	asset: Asset;
-	_type: string;
-	alt: string;
-}
-
-export interface Asset {
-	_ref: string;
-	_type: string;
-}
