+++ app/src/routes/statement-of-faith/+page.ts
@@ -6,7 +6,7 @@ export const prerender = true;
 export const load: PageLoad = async ({ url }) => {
 	const { pathname } = url;
 
-	const data= await S.getPageData(pathname);
+	const data= await S.getPage(pathname);
 
 	return {
 		page: data
